refactor: phone number input into its own component
